chore(Header): Create a desktop header and a mobile header (#370)
