+++ sensemap/src/components/Header/DesktopHeader/Breadcrumb/index.tsx
@@ -2,13 +2,13 @@ import * as React from 'react';
 import { Switch, Route, Link, withRouter, RouteComponentProps } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { Breadcrumb as SBreadcrumb } from 'semantic-ui-react';
-import * as T from '../../../types';
-import { actions, ActionProps, mapDispatch } from '../../../types';
-import * as B from '../../../types/sense/box';
-import * as M from '../../../types/sense/map';
-import { MapScope } from '../../../types/sense-map';
-import * as CS from '../../../types/cached-storage';
-import * as R from '../../../types/routes';
+import * as T from '../../../../types';
+import { actions, ActionProps, mapDispatch } from '../../../../types';
+import * as B from '../../../../types/sense/box';
+import * as M from '../../../../types/sense/map';
+import { MapScope } from '../../../../types/sense-map';
+import * as CS from '../../../../types/cached-storage';
+import * as R from '../../../../types/routes';
 
 // tslint:disable-next-line:no-any
 type MyRouteProps = RouteComponentProps<any>;
